#18: Added a DiscogsCollectionImporter, which right now is just a file with extensions to the managed objects for initialization with their corresponding Discogs objects.

+++ SwiftDiscogsApp/Third-Party Services/Discogs/DiscogsService.swift
@@ -225,7 +225,7 @@ class DiscogsService: ThirdPartyService, AuthenticatedService, ImportableService
         do {
             let coreDataItem = try fetchOrCreateItem(releaseVersionID: item.id,
                                                      inContext: context)
-            coreDataItem.rating = Int64(item.rating)
+            coreDataItem.rating = Int16(item.rating)
             folder.addToItems(coreDataItem)
             importedItemCount += 1
         } catch {
